third_party/blink/web_tests/svg/dynamic-updates/SVGFEDropShadowElement-dom-stdDeviation-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stdDeviation' attribute of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dropShadowElement.getAttribute('stdDeviation') is "0"
PASS dropShadowElement.getAttribute('stdDeviation') is "10"
PASS successfullyParsed is true

TEST COMPLETE
